refactor!: split API module into separate files for better organization
Split the monolithic api.py file into multiple modules:
- client.py: Contains the main API class with all methods
- types.py: Contains all type definitions, enums, and constants
- errors.py: Contains all custom exception classes
- __init__.py: Provides backwards-compatible imports

This change improves code maintainability and readability by separating concerns. The API interface remains the same with all classes and functions re-exported from the api module.

BREAKING CHANGE: Internal import paths have changed. Code importing specific classes directly from pythreads.api.Field, pythreads.api.MediaType, etc. may break. Use imports from pythreads.api instead.
