+++ src/pythreads/api/__init__.py
@@ -1,0 +1,38 @@
+from .client import API
+from .types import (
+    Field,
+    Media,
+    ContainerStatus,
+    MediaType,
+    ReplyControl,
+    PublishingStatus,
+    PublishingError,
+)
+from .errors import (
+    ThreadsInvalidParameter,
+    ThreadsResponseError,
+    ThreadsHTTPError,
+)
+
+# Re-export Threads for test patching compatibility: tests patch
+# "pythreads.api.Threads.build_graph_api_url".
+from pythreads.threads import Threads  # noqa: E402,F401
+
+__all__ = [
+    "API",
+    # Types
+    "Field",
+    "Media",
+    "ContainerStatus",
+    "MediaType",
+    "ReplyControl",
+    "PublishingStatus",
+    "PublishingError",
+    # Errors
+    "ThreadsInvalidParameter",
+    "ThreadsResponseError",
+    "ThreadsHTTPError",
+    # Compatibility
+    "Threads",
+]
+
